feat(myparcel): add shipment strategy
