+++ packages/vendure-plugin-myparcel/src/api/types.ts
@@ -1,5 +1,8 @@
-import { OrderLine } from '@vendure/core';
+import { Order, OrderLine } from '@vendure/core';
+import { OrderAddress } from '@vendure/common/lib/generated-types';
+import { GraphQLError } from 'graphql';
 
+export type PartialBy<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
 export interface MyparcelConfig {
   vendureHost: string;
   /**
@@ -11,6 +14,7 @@ export interface MyparcelConfig {
    * {@link CustomsInformation} per order item
    */
   getCustomsInformationFn?: (orderLine: OrderLine) => CustomsInformation;
+  shipmentStrategy: MyParcelShipmentStrategy;
 }
 
 /**
@@ -30,4 +34,78 @@ export interface CustomsInformation {
    * Country of origin. Has to be a capitalized country code, I.E. 'NL'
    */
   countryCodeOfOrigin: string;
+}
+
+export interface MyparcelRecipient {
+  cc: string;
+  region?: string;
+  city: string;
+  street: string;
+  number: string;
+  number_suffix?: string;
+  postal_code: string;
+  person: string;
+  phone?: string;
+  email?: string;
+}
+
+export interface MyparcelShipmentOptions {
+  package_type: number;
+  label_description?: string;
+}
+
+export interface CustomsDeclaration {
+  contents: string | number;
+  invoice: string;
+  weight: number;
+  items: CustomsItem[];
+}
+
+export interface ItemValue {
+  amount: number;
+  currency: string;
+}
+
+export interface CustomsItem {
+  description: string;
+  amount: number;
+  weight: number;
+  item_value: ItemValue;
+  classification: string;
+  country: string;
+}
+
+export interface MyparcelShipment {
+  carrier: number;
+  reference_identifier?: string;
+  recipient: MyparcelRecipient;
+  options: MyparcelShipmentOptions;
+  customs_declaration?: CustomsDeclaration;
+  physical_properties?: {
+    weight: number;
+  };
+}
+
+export interface MyParcelShipmentStrategy {
+  getOptions: (
+    address: OrderAddress,
+    order: Order,
+    customsContent: string
+  ) => MyparcelShipment['options'];
+  getRecipient: (
+    address: OrderAddress,
+    order: Order,
+    customsContent: string
+  ) => MyparcelShipment['recipient'];
+  getShipment: (
+    address: OrderAddress,
+    order: Order,
+    customsContent: string
+  ) => MyparcelShipment;
+}
+
+export class MyParcelError extends GraphQLError {
+  constructor(message: string) {
+    super(message, { extensions: { code: 'MY_PARCEL_ERROR' } });
+  }
 }
